Workflows: tests fixes * Use dag.test() in order to quickly test the whole dag run * Separated dag and harvesting methods tests
