+++ airflow.cfg
@@ -110,7 +110,7 @@ load_default_connections = True
 execute_tasks_new_python_interpreter = False
 
 # Secret key to save connection passwords in the db
-fernet_key = 
+fernet_key =
 
 # Whether to disable pickling dags
 donot_pickle = True
@@ -392,7 +392,7 @@ enable_experimental_api = False
 # ("airflow.api.auth.backend.default" allows all requests for historic reasons)
 auth_backends = airflow.api.auth.backend.basic_auth
 # has had airflow.api.auth.backend.session added in the running config, which is needed by the UI. Please update your config before Apache Airflow 3.0.
-# auth_backends = airflow.api.auth.backend.session 
+# auth_backends = airflow.api.auth.backend.session
 # Used to set the maximum page limit for API requests
 maximum_page_limit = 100
 
@@ -424,7 +424,7 @@ access_control_allow_methods = POST, GET, OPTIONS, DELETE
 
 # Indicates whether the response can be shared with requesting code from the given origins.
 # Separate URLs with space.
-access_control_allow_origins = 
+access_control_allow_origins =
 
 [lineage]
 # what lineage backend to use
