+++ dags/oup/ftp_service.py
@@ -5,10 +5,6 @@ from common.ftp_service import FTPService
 
 class OUPFTPService(FTPService):
     def __init__(self):
-        super().__init__(
-            host=os.getenv("OUP_FTP_HOST", "localhost"),
-            username=os.getenv("OUP_FTP_USERNAME", "airflow"),
-            password=os.getenv("OUP_FTP_PASSWORD", "airflow"),
-            port=int(os.getenv("OUP_FTP_PORT", "2222")),
-            dir=os.getenv("OUP_FTP_DIR", "upload/oup"),
-        )
+        super().__init__()
+        self.ftp_conn_id = "oup_ftp_service"
+        self.dir = os.getenv("OUP_FTP_DIR", "/upload/oup")
