+++ requirements-airflow.txt
@@ -1,5 +1,6 @@
-apache-airflow[celery]==2.6.0 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.6.0/constraints-3.8.txt
-apache-airflow-providers-celery==2.1.3
+apache-airflow-providers-celery==3.1.0
+apache-airflow-providers-ftp==3.3.1
 celery==5.2.6
 redis==4.2.2
 psycopg2==2.9.6
+apache-airflow[celery]==2.6.1 --constraint https://raw.githubusercontent.com/apache/airflow/constraints-2.6.1/constraints-3.10.txt
