+++ tests/units/hindawi/data/example1.xml
@@ -66,4 +66,3 @@
 		</ns1:record>
 	</ns0:metadata>
 </ns0:record>
-
