+++ tests/units/hindawi/data/example2.xml
@@ -182,4 +182,3 @@
 		</ns1:record>
 	</ns0:metadata>
 </ns0:record>
-
