+++ tests/units/hindawi/data/without_doi.xml
@@ -55,4 +55,3 @@
 		</ns1:record>
 	</ns0:metadata>
 </ns0:record>
-
