+++ tests/units/hindawi/data/without_page_nr.xml
@@ -181,4 +181,3 @@
 		</ns1:record>
 	</ns0:metadata>
 </ns0:record>
-
