+++ tests/units/oup/data/ptab170_no_journal_title_and_pubmed_value.xml
@@ -37,7 +37,7 @@
                 </subj-group>
             </article-categories>
             <title-group>
-                <article-title>Detectable electric current induced by the dark matter axion in a 
+                <article-title>Detectable electric current induced by the dark matter axion in a
                     conductor</article-title>
             </title-group>
             <contrib-group>
